feat/RSS-ECOMM-2_24: appropriate lazy roating
